backend/service_auth/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
pydantic==2.9.2
python-multipart==0.0.12
email-validator==2.1.0

